requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
diff-match-patch==20230430
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor-5==0.2.12
django-import-export==3.3.7
django-jazzmin==3.0.0
django-redis==3.1.6
django-timezone-field==6.1.0
flower==2.0.1
humanize==4.9.0
kombu==5.3.7
MarkupPy==1.14
openpyxl==3.1.2
pillow==10.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0

python-dotenv~=1.0.1